+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ---
 
+## Master
+
+### Bug fixes
+
+- Fixed parsing protocol method return type followed by declaration with attribute
+
+
 ## 0.12.0
 
 ### Internal Changes