+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Bug fixes
 
 - Fixed parsing protocol method return type followed by declaration with attribute
+- Fixed inserting auto-inlined code on the last line of declaration body
 
 
 ## 0.12.0