+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Fixed parsing protocol method return type followed by declaration with attribute
 - Fixed inserting auto-inlined code on the last line of declaration body
+- AutoEquatable and AutoHashable templates should not add protocol conformances in extensions
 
 ## 0.12.0