+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ---
 
-## Master
+## 0.13.0
 
 ### New Features